backend/requirements.txt
Django==4.2.3
django-environ==0.9.0
psycopg2-binary
python-dotenv
djangorestframework
django-cors-headers
djangorestframework-simplejwt==4.7.2
requests
pyjwt

